refactor: split functions across files and re-export
